#781: (Chrome) zoom handling
Chrome changes windows.devicePixelRatio when zooming on browser level.
Improved behaviour by changing property type to Double to accommodate for
ratios like 0.8 produced by Chrome, but it's still not perfect.

+++ draw/jvm/src/main/ch/scorpion/jabbah/draw/view/CanvasJvm.kt
@@ -86,7 +86,7 @@ class CanvasJvm(
 		background = Graphics2DJvm.toAwtColor(styleProvider.getStyle(StyleType.BACKGROUND).color.backgroundColor)
 	}
 
-	override val devicePixelRatio: Int = 1
+	override val devicePixelRatio: Double = 1.0
 
 	override var backgroundColor: Color
 		get() = Color(background.red, background.green, backgroundColor.blue, backgroundColor.alpha)
